fix: fetchdata

+++ src/libs/semesters/deleteSemester.tsx
@@ -1,18 +1,25 @@
-export default async function deleteSemester(did: number) {
+export default async function deleteSemester(did: number): Promise<any> {
   try {
-    const response = await fetch(
-      `${process.env.PUBLIC_BACKEND_URL}api/v1/semesters/${did}`,
-      {
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${process.env.USER_TOKEN}`,
-        },
-      }
-    )
-    if (!response.ok) {
-      throw new Error('Failed to delete semester')
+    const backendUrl = process.env.PUBLIC_BACKEND_URL
+    const userToken = process.env.USER_TOKEN
+
+    if (!backendUrl || !userToken) {
+      throw new Error('Environment variables are not set correctly')
     }
+
+    const response = await fetch(`${backendUrl}api/v1/semesters/${did}`, {
+      method: 'DELETE',
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: `Bearer ${userToken}`,
+      },
+    })
+
+    if (!response.ok) {
+      const errorText = await response.text()
+      throw new Error(`Failed to delete semester: ${errorText}`)
+    }
+
     return await response.json()
   } catch (err) {
     console.error('Error deleting semester:', err)
